[monorepo] Remove accidentally added top-level monorepo directory
Apparently I introduced a new top level directory called `b` in the root
of the monorepo with r353906. I don't understand how or why this
happened, but I presume something went wrong doing `git llvm push`.

I'm not even sure if I can delete the directory again, but giving it a
shot in this commit.

Thanks to Shoaib Meenai for noticing this.

git-svn-id: https://llvm.org/svn/llvm-project/monorepo-root/trunk@353913 91177308-0d34-0410-b5e6-96231b3b80d8